move test to color_test.go
